requirements.txt
fastapi[standard]>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
azure-storage-blob>=12.23.0
openai>=1.55.0
python-docx>=1.1.0
PyMuPDF>=1.24.0
pydantic>=2.9.0
requests>=2.32.0
numpy<2.0.0
docling>=2.10.0
sentence-transformers>=3.0.0
plotly>=5.24.0

# Audience Generation dependencies
langchain-openai>=0.2.0
langchain-core>=0.3.0
